Implemented edge case handling for Alexandria
Implemented fixes for the following:

Checking for edge cases and properly handling errors
-The application crashed when there is no internet connection instead of
providing an appropriate error message.
-When internet connection is missing, the download of the book image
fails and the application does not provide a proper placeholder for the
missing image.

Accessibility

The application does not implement accessibility features.
- There are some missing content descriptions for user interactable
controls in any of the user interfaces.
Directional Pad or keyboard navigation for the interface elements are
not handled explicitly which can be a problem as the app grows.

Missing functionality

- The scan functionality is not implemented yet.

Localization

Localizable resources should be marked as localizable while resources
that stay the same in any language (like setting keys for example)
should be marked as non localizable.